+++ src/pycex/__init__.py
@@ -4,6 +4,7 @@ from pycex.base import BaseExchange
 from pycex.constants import (
     BINANCE_BASE,
     BINANCE_TESTNET,
+    BITGET_BASE,
     BUY,
     BYBIT_BASE,
     BYBIT_TESTNET,
@@ -23,16 +24,18 @@ from pycex.exceptions import (
     RateLimitError,
     SymbolNotFoundError,
 )
-from pycex.exchanges import OKX, Binance, Bybit
+from pycex.exchanges import OKX, Binance, Bitget, Bybit
 
 __version__ = "0.1.0"
 __all__ = [
     "BaseExchange",
     "Binance",
+    "Bitget",
     "Bybit",
     "OKX",
     "BINANCE_BASE",
     "BINANCE_TESTNET",
+    "BITGET_BASE",
     "BYBIT_BASE",
     "BYBIT_TESTNET",
     "OKX_BASE",