+++ src/pycex/exchanges/__init__.py
@@ -1,7 +1,8 @@
 """Exchange implementations."""
 
 from pycex.exchanges.binance import Binance
+from pycex.exchanges.bitget import Bitget
 from pycex.exchanges.bybit import Bybit
 from pycex.exchanges.okx import OKX
 
-__all__ = ["Binance", "Bybit", "OKX"]
+__all__ = ["Binance", "Bitget", "Bybit", "OKX"]